Move Inputs definition to parameter.py
