+++ pybop/costs/_likelihoods.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pybop.costs.base_cost import BaseCost
-from pybop.models.base_model import Inputs
+from pybop.parameters.parameter import Inputs
 
 
 class BaseLikelihood(BaseCost):
