+++ pybop/costs/design_costs.py
@@ -3,7 +3,7 @@ import warnings
 import numpy as np
 
 from pybop.costs.base_cost import BaseCost
-from pybop.models.base_model import Inputs
+from pybop.parameters.parameter import Inputs
 
 
 class DesignCost(BaseCost):
