+++ pybop/costs/fitting_costs.py
@@ -2,8 +2,8 @@ import numpy as np
 
 from pybop.costs._likelihoods import BaseLikelihood
 from pybop.costs.base_cost import BaseCost
-from pybop.models.base_model import Inputs
 from pybop.observers.observer import Observer
+from pybop.parameters.parameter import Inputs
 
 
 class RootMeanSquaredError(BaseCost):
