+++ pybop/models/empirical/ecm.py
@@ -1,7 +1,7 @@
 from pybamm import equivalent_circuit as pybamm_equivalent_circuit
 
-from pybop.models.base_model import Inputs
 from pybop.models.empirical.base_ecm import ECircuitModel
+from pybop.parameters.parameter import Inputs
 
 
 class Thevenin(ECircuitModel):
