+++ pybop/problems/fitting_problem.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from pybop import BaseProblem
-from pybop.models.base_model import Inputs
+from pybop.parameters.parameter import Inputs
 
 
 class FittingProblem(BaseProblem):
